Add verifications if file exists, is regular, writable (and readable) in FileWriter

+++ src/util/writer/FileWriter.java
@@ -1,21 +1,21 @@
 package util.writer;
 
-import model.student.Student;
-import util.reader.FileReader;
-
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 public class FileWriter {
     public boolean appendLine(String line, Path path) {
         try {
-            Files.write(path, line.getBytes(), StandardOpenOption.APPEND);
-            return true;
+            if (Files.exists(path) && Files.isRegularFile(path)
+                    && Files.isWritable(path)) {
+                Files.write(path, line.getBytes(), StandardOpenOption.APPEND);
+                return true;
+            }
+            return false;
         } catch (IOException e) {
             System.out.println("File not found." + e.getMessage());
             return false;
@@ -27,8 +27,12 @@ public class FileWriter {
 
     public boolean deleteLine(String line, Path path){
         try {
-            Files.write(path, Files.lines(path).filter(fileLine -> fileLine.contentEquals(line) ? false : true).collect(Collectors.toList()));
-            return true;
+            if (Files.exists(path) && Files.isRegularFile(path)
+                    && Files.isReadable(path) && Files.isWritable(path)) {
+                Files.write(path, Files.lines(path).filter(fileLine -> fileLine.contentEquals(line) ? false : true).collect(Collectors.toList()));
+                return true;
+            }
+            return false;
         } catch (NullPointerException exception) {
             System.out.println("Illegal argument. Line cannot be null");
             return false;
@@ -40,10 +44,14 @@ public class FileWriter {
 
     public boolean rewriteLine(String line, int lineNumber, Path path){
         try {
-            List<String> text = Files.readAllLines(path);
-            text.set(lineNumber, line);
-            Files.write(path, text);
-            return true;
+            if (Files.exists(path) && Files.isRegularFile(path)
+                    && Files.isReadable(path) && Files.isWritable(path)) {
+                List<String> text = Files.readAllLines(path);
+                text.set(lineNumber, line);
+                Files.write(path, text);
+                return true;
+            }
+            return false;
         } catch (IOException e) {
             System.out.println("File not found");
             return false;
